requirements.txt
google-adk==1.18.0
google-genai==0.12.0
pandas==2.1.4
scikit-learn==1.4.0
tqdm==4.66.1
tabulate==0.9.0
pytest==8.4.2
pytest-asyncio==0.23.2